Make ProcessId obligatory on Ticket and update more relations
Modified `ProcessId` and `WorkflowId` data types to `long` for consistency. Adjusted entity relationships to enforce required constraints, ensuring `Process` and `Workflow` associations are mandatory where applicable.

+++ o2rabbit.Core/Entities/Process.cs
@@ -18,7 +18,7 @@ public class Process
 
     public List<Ticket> Tickets { get; } = [];
 
-    public int WorkflowId { get; set; }
+    public long WorkflowId { get; set; }
 
     public Workflow? Workflow { get; set; }
 }
